Minor refactor to improve organization
This moves the tile reader into a separate type
and separate file, and removes some debugging
leftovers.
